+++ TiendaComputadoras/src/logica/DiscoDuro.java
@@ -4,7 +4,7 @@ public class DiscoDuro extends Componente {
 	private String CapacidadAlma;
 	private String tipoConexion;
 	
-	public DiscoDuro(String modelo, String numeroSerie, String marca, int cantDisponible, float cantMin, float cantMax,
+	public DiscoDuro(String modelo, String numeroSerie, String marca, int cantDisponible, int cantMin, int cantMax,
 			float precioVentaI, float precioCompraI, String capacidadAlma, String tipoConexion) {
 		super(modelo, numeroSerie, marca, cantDisponible, cantMin, cantMax, precioVentaI, precioCompraI);
 		CapacidadAlma = capacidadAlma;
