+++ TiendaComputadoras/src/logica/MotherBoard.java
@@ -5,8 +5,8 @@ public class MotherBoard extends Componente {
 	private String tipoRAM;
 	//private ArrayList<String> 
 	
-	public MotherBoard(String modelo, String numeroSerie, String marca, int cantDisponible, float cantMin,
-			float cantMax, float precioVentaI, float precioCompraI, String tipoConector, String tipoRAM) {
+	public MotherBoard(String modelo, String numeroSerie, String marca, int cantDisponible, int cantMin,
+			int cantMax, float precioVentaI, float precioCompraI, String tipoConector, String tipoRAM) {
 		super(modelo, numeroSerie, marca, cantDisponible, cantMin, cantMax, precioVentaI, precioCompraI);
 		this.tipoConector = tipoConector;
 		this.tipoRAM = tipoRAM;
